fix: DRINK_VOLUME aus use server Datei in lib auslagern
